+++ .dialyzer_ignore.exs
@@ -1,0 +1,3 @@
+[
+  {"lib/ref_inspector/config.ex", :contract_supertype}
+]