+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v1.1.0-dev
 
 - Enhancements
+    - Default configuration entries for files and urls are available through `RefInspector.Config.default_files/0` and `RefInspector.Config.default_urls/0`
     - Output of mix task `ref_inspector.download` can be prevented by passing `--quiet` upon invocation. This does NOT imply `--force` and will still ask for confirmation
     - The library used to read YAML files can be changed by using the `:yaml_file_reader` configuration